# backend/openapi/partner-v1.yaml
openapi: 3.0.3
info:
  title: CPS Partner API
  version: "1.0.0"
  contact:
    name: CPS Logistics
    url: https://cpslogistics.com/contact
  description: |
    The CPS Partner API lets businesses (marketplaces, shops, platforms) send
    deliveries to **CPS Logistics** programmatically, charge their customers
    CPS's price, and follow every delivery live.

    # Overview

    A typical integration:

    1. **Quote** — at checkout, call `POST /quotes` with the pickup and
       drop-off regions. Show the returned `fee` to your customer.
    2. **Create** — when the order is ready to collect, call
       `POST /shipments` with your own order id as `externalReference` and the
       quoted fee as `expectedFee`. CPS dispatches a rider.
    3. **Follow** — CPS calls your **webhook** whenever the delivery changes
       (rider assigned, picked up, out for delivery, delivered…). You can also
       poll `GET /shipments/{ref}`.
    4. **Settle** — pay CPS for delivered orders. Your Business portal has a
       statement of every delivery and fee for any period.

    CPS is based in **Kumasi**. Riders collect from the pickup regions listed by
    `GET /coverage` and deliver to the drop-off regions listed there.

    # Getting access

    1. Sign up at CPS and choose **Business (API access)** as the account type.
    2. CPS reviews and approves your business (you'll get a notification).
    3. In the Business portal open **API keys → Create key**. Copy the key —
...

...

...

    ```
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /coverage:
    get:
      tags: [Coverage & pricing]
      operationId: getCoverage
      summary: Regions, fees and operating hours
      description: The live rate card. Cache it for a few hours at most; rates can change.
      responses:
        "200":
          description: Coverage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Coverage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /quotes:
    post:
      tags: [Coverage & pricing]
      operationId: createQuote
      summary: Price a delivery
      description: Returns the binding CPS fee for a route, or why CPS can't serve it. Region names are matched case-insensitively.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuoteRequest" }
            examples:
              accra:
                summary: Kumasi to Accra
                value: { pickupRegion: Kumasi, dropoffRegion: Accra }
              campus:
                summary: Within Kumasi, KNUST area
                value: { pickupRegion: Kumasi, dropoffRegion: Kumasi, dropoffKumasiSubArea: CampusAndEnvirons }
      responses:
        "200":
          description: Quote (check `serviceable`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quote" }
              examples:
                serviceable:
                  value: { serviceable: true, fee: 45, currency: GHS, pickupRegion: Kumasi, dropoffRegion: Accra }
                notServiceable:
                  value: { serviceable: false, reason: pickup_region_not_served, message: 'CPS does not collect from "Accra". Pickup regions: Kumasi.' }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /shipments:
    post:
      tags: [Shipments]
      operationId: createShipment
      summary: Create a shipment
      description: |
        Books a delivery. Idempotent on `externalReference`: repeating a create
        returns the original shipment with `200`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateShipmentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shipment" }
        "200":
          description: Already existed (same `externalReference`); the original is returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shipment" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`fee_mismatch` — the current fee differs from `expectedFee`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: { code: fee_mismatch, message: "The delivery fee is GHS 45.00, not GHS 30.00. Re-quote and retry.", details: { fee: 45 } } }
        "422":
          description: "`not_serviceable` — CPS doesn't serve this route"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
    get:
      tags: [Shipments]
      operationId: listShipments
      summary: List shipments
      description: |
        Your shipments, **least recently updated first**, for reconciliation.
        To sync, remember the `updatedAt` of the last shipment you processed and
        pass it as `updatedSince`; page with `cursor`.
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ShipmentStatus" }
        - name: updatedSince
          in: query
          description: Only shipments updated at or after this time (ISO 8601).
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          description: "`nextCursor` from the previous page."
          schema: { type: string }
      responses:
        "200":
          description: A page of shipments
          content:
            application/json:
              schema:
                type: object
                required: [data, nextCursor]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Shipment" }
                  nextCursor:
                    type: string
                    nullable: true
                    description: Pass as `cursor` for the next page; `null` on the last page.
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /shipments/{ref}:
    get:
      tags: [Shipments]
      operationId: getShipment
      summary: Get a shipment
      parameters:
        - $ref: "#/components/parameters/ShipmentRef"
      responses:
        "200":
          description: The shipment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shipment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /shipments/{ref}/cancel:
    post:
      tags: [Shipments]
      operationId: cancelShipment
      summary: Cancel a shipment
      description: |
        Only possible before the rider picks the package up (status `pending`).
        Cancelling an already-cancelled shipment returns it unchanged. After
        pickup, contact CPS operations.
      parameters:
        - $ref: "#/components/parameters/ShipmentRef"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 500, example: Customer cancelled the order }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shipment" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`conflict` — already picked up or finished"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "cps_live_…"
      description: Secret API key from the CPS Business portal.
  parameters:
    ShipmentRef:
      name: ref
      in: path
      required: true
      description: Your `externalReference`, or the CPS `trackingCode`.
      schema: { type: string }
      example: order_10293
  responses:
    InvalidRequest:
      description: "`invalid_request`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: invalid_request, message: "deliveryCode: must be 4 to 8 digits", details: [{ path: deliveryCode, message: must be 4 to 8 digits }] } }
    Unauthorized:
      description: "`unauthorized` — missing, malformed or revoked key"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: "`forbidden` — business pending approval or suspended"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: "`not_found`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: "`rate_limited`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, unauthorized, forbidden, not_found, fee_mismatch, conflict, not_serviceable, rate_limited, internal_error]
            message: { type: string }
            details: { description: Extra machine-readable context; shape depends on `code`. }
    Region:
      type: string
      description: Region name as listed by `GET /coverage` (case-insensitive).
      example: Accra
    KumasiSubArea:
      type: string
      enum: [CampusAndEnvirons, Other]
      description: |
        Only for drop-offs in Kumasi. `CampusAndEnvirons` = KNUST campus,
        Ayeduase, Ayigya, Bomso, Kotei. Defaults to `Other`.
    ShipmentStatus:
      type: string
      enum: [pending, picked_up, in_transit, out_for_delivery, delayed, delivered, failed, cancelled]
      description: See "Shipment lifecycle". New values may be added; treat unknown values as "in progress".
    Coverage:
      type: object
      required: [currency, pickupRegions, dropoffRegions, operatingHours]
      properties:
        currency: { type: string, example: GHS }
        pickupRegions:
          type: array
          items: { $ref: "#/components/schemas/Region" }
          example: [Kumasi]
        dropoffRegions:
          type: array
          items:
            type: object
            required: [region, fee]
            properties:
              region: { $ref: "#/components/schemas/Region" }
              kumasiSubArea:
                allOf: [{ $ref: "#/components/schemas/KumasiSubArea" }]
                nullable: true
              description: { type: string, nullable: true }
              fee: { type: number, example: 45 }
        operatingHours:
          type: object
          properties:
            timezone: { type: string, example: Africa/Accra }
            api:
              type: object
              properties:
                days: { type: string, example: Mon-Sun }
                open: { type: string, example: "05:00" }
                close: { type: string, example: "20:00" }
            riders:
              type: object
              properties:
                days: { type: string, example: Mon-Sat }
                open: { type: string, example: "08:00" }
                close: { type: string, example: "19:30" }
    QuoteRequest:
      type: object
      required: [pickupRegion, dropoffRegion]
      properties:
        pickupRegion: { $ref: "#/components/schemas/Region" }
        dropoffRegion: { $ref: "#/components/schemas/Region" }
        dropoffKumasiSubArea: { $ref: "#/components/schemas/KumasiSubArea" }
    Quote:
      oneOf:
        - type: object
          title: Serviceable
          required: [serviceable, fee, currency, pickupRegion, dropoffRegion]
          properties:
            serviceable: { type: boolean, enum: [true] }
            fee: { type: number, example: 45 }
            currency: { type: string, example: GHS }
            pickupRegion: { $ref: "#/components/schemas/Region" }
            dropoffRegion: { $ref: "#/components/schemas/Region" }
        - type: object
          title: Not serviceable
          required: [serviceable, reason, message]
          properties:
            serviceable: { type: boolean, enum: [false] }
            reason: { type: string, enum: [pickup_region_not_served, dropoff_region_not_served] }
            message: { type: string }
    Contact:
      type: object
      required: [name, phone, region, location]
      properties:
        name: { type: string, maxLength: 120, example: Ama Store }
        phone: { type: string, minLength: 7, maxLength: 20, example: "0241234567" }
        region: { $ref: "#/components/schemas/Region" }
        location:
          type: string
          maxLength: 300
          description: Street address and landmarks the rider can find.
          example: Adum, Shop 12, near Kejetia
    CreateShipmentRequest:
      type: object
      required: [externalReference, pickup, dropoff, expectedFee]
      properties:
        externalReference:
          type: string
          maxLength: 100
          description: Your id for this delivery (e.g. your order id). Unique per business; makes creates idempotent.
          example: order_10293
        pickup: { $ref: "#/components/schemas/Contact" }
        dropoff:
          allOf:
            - $ref: "#/components/schemas/Contact"
            - type: object
              properties:
                kumasiSubArea: { $ref: "#/components/schemas/KumasiSubArea" }
                latitude: { type: number, minimum: -90, maximum: 90, description: Optional; gives the rider a map link. }
                longitude: { type: number, minimum: -180, maximum: 180 }
        package:
          type: object
          properties:
            type: { type: string, enum: [document, parcel, electronics, fragile, food, other], default: parcel }
            size: { type: string, enum: [small, medium, big], default: medium }
            weightKg: { type: number, exclusiveMinimum: true, minimum: 0 }
            description: { type: string, maxLength: 300, example: "2 pairs of shoes" }
        speed: { type: string, enum: [same_day, next_day, express], default: next_day }
        priority: { type: string, enum: [standard, high], default: standard }
        expectedFee:
          type: number
          description: The fee you quoted to your customer. Must equal CPS's current fee, or the request fails with `fee_mismatch`.
          example: 45
        deliveryCode:
          type: string
          pattern: "^\\d{4,8}$"
          description: Code your customer must give the rider to receive the package. See "Delivery codes".
          example: "482913"
        instructions: { type: string, maxLength: 1000, example: Call on arrival; gate is blue. }
    RiderContact:
      type: object
      nullable: true
      properties:
        name: { type: string, example: Yaw Boateng }
        phone: { type: string, nullable: true, example: "0551234567" }
    Shipment:
      type: object
      required: [trackingCode, externalReference, status, fee, pickup, dropoff, package, rider, riders, deliveryCodeRequired, proofOfDelivery, cancellationReason, events, createdAt, updatedAt]
      properties:
        trackingCode: { type: string, example: CPS-7KQ2M9XA, description: CPS's tracking code. }
        externalReference: { type: string, nullable: true, example: order_10293 }
        status: { $ref: "#/components/schemas/ShipmentStatus" }
        deliveryType: { type: string, enum: [doorstep, station] }
        fee:
          type: object
          properties:
            amount: { type: number, example: 45 }
            currency: { type: string, example: GHS }
        pickup: { $ref: "#/components/schemas/Contact" }
        dropoff:
          allOf:
            - $ref: "#/components/schemas/Contact"
            - type: object
              properties:
                kumasiSubArea:
                  allOf: [{ $ref: "#/components/schemas/KumasiSubArea" }]
                  nullable: true
        package:
          type: object
          properties:
            type: { type: string }
            size: { type: string }
        rider:
          allOf: [{ $ref: "#/components/schemas/RiderContact" }]
          description: The rider currently responsible — pickup rider until pickup, then the delivery rider. `null` until assigned.
        riders:
          type: object
          properties:
            pickup: { $ref: "#/components/schemas/RiderContact" }
            dropoff: { $ref: "#/components/schemas/RiderContact" }
        deliveryCodeRequired: { type: boolean }
        proofOfDelivery:
          type: object
          nullable: true
          description: Set once `status` is `delivered`.
          properties:
            method: { type: string, enum: [signature, photo], nullable: true }
            recipientName: { type: string, nullable: true }
            photoUrl: { type: string, nullable: true }
            deliveryCodeVerified: { type: boolean }
        cancellationReason: { type: string, nullable: true }
        events:
          type: array
          description: Status history, oldest first.
          items:
            type: object
            properties:
              status: { $ref: "#/components/schemas/ShipmentStatus" }
              note: { type: string, nullable: true }
              at: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...